Initial logical volumes (LVM) As well as encryption support.
- Support for logical volumes so we can get access to a larger /system
partition by taking space from /data (mmcblk0p2) /system initial size is
600MB and can get easily upgradable by tweaking the updater.sh

- Adding support as well for encryption. Depending when data/data is encrypted
or not, the contents of it goes either on data/data or on the /datadata
partition.

- Updated the updater.sh from aries to take care of the LVM handling.

Initially made for Aries devices, adapted over for YP-G1.

+++ updater.sh
@@ -2,15 +2,10 @@
 #
 # Universal Updater Script for Samsung Galaxy S Phones
 # (c) 2011 by Teamhacksung
-# GSM version
+# Combined GSM & CDMA version, Adapted for YP-G1 (Samsung Galaxy Player/S Wifi 4.0)
 #
 
-set -x
-export PATH=/:/sbin:/system/xbin:/system/bin:/tmp:$PATH
-
-# check if we're running on a bml or mtd device
-if /tmp/busybox test -e /dev/block/bml7 || [ $(grep mtdblock2 /proc/partitions | awk '{ print $3 }') -lt 395264 ]; then
-# we're running on a bml device or using an older mtd partition layout
+SYSTEM_SIZE='629145600' # 600M
 
 check_mount() {
     local MOUNT_POINT=`/tmp/busybox readlink $1`
@@ -35,9 +30,36 @@ set_log() {
     exec >> $1 2>&1
 }
 
+warn_repartition() {
+    if ! /tmp/busybox test -e /.accept_wipe ; then
+        /tmp/busybox touch /.accept_wipe
+        ui_print
+        ui_print "============================================"
+        ui_print "This ROM uses an incompatible partition layout"
+        ui_print "Your /data will be wiped upon installation"
+        ui_print "Run this update.zip again to confirm install"
+        ui_print "============================================"
+        ui_print
+        exit 9
+    fi
+    /tmp/busybox rm /.accept_wipe
+}
+
+format_partitions() {
+    /lvm/sbin/lvm lvcreate -L ${SYSTEM_SIZE}B -n system lvpool
+    /lvm/sbin/lvm lvcreate -l 100%FREE -n userdata lvpool
+
+    # format data (/system will be formatted by updater-script)
+    /tmp/make_ext4fs -b 4096 -g 32768 -i 8192 -I 256 -l -16384 -a /data /dev/lvpool/userdata
+
+    # unmount and format datadata
+    /tmp/busybox umount -l /datadata
+    /tmp/erase_image datadata
+}
+
 fix_package_location() {
     local PACKAGE_LOCATION=$1
-    # Remove leading /mnt
+    # Remove leading /mnt for Samsung recovery
     PACKAGE_LOCATION=${PACKAGE_LOCATION#/mnt}
     # Convert to modern sdcard path
     PACKAGE_LOCATION=`echo $PACKAGE_LOCATION | /tmp/busybox sed -e "s|^/sdcard|/storage/sdcard0|"`
@@ -45,27 +67,68 @@ fix_package_location() {
     echo $PACKAGE_LOCATION
 }
 
+# ui_print by Chainfire
+OUTFD=$(/tmp/busybox ps | /tmp/busybox grep -v "grep" | /tmp/busybox grep -o -E "update_binary(.*)" | /tmp/busybox cut -d " " -f 3);
+ui_print() {
+  if [ $OUTFD != "" ]; then
+    echo "ui_print ${1} " 1>&$OUTFD;
+    echo "ui_print " 1>&$OUTFD;
+  else
+    echo "${1}";
+  fi;
+}
+
+set -x
+export PATH=/:/sbin:/system/xbin:/system/bin:/tmp:$PATH
+
+# Check if we're in CDMA or GSM mode
+if /tmp/busybox test "$1" = cdma ; then
+    # CDMA mode
+    IS_GSM='/tmp/busybox false'
+    SD_PART='/dev/block/mmcblk1p1'
+    MMC_PART='/dev/block/mmcblk0p1 /dev/block/mmcblk0p2'
+    MTD_SIZE='490733568'
+else
+    # GSM mode/Wi-fi mode
+    IS_GSM='/tmp/busybox true'
+    SD_PART='/dev/block/mmcblk0p1'
+    MMC_PART='/dev/block/mmcblk0p2'
+    MTD_SIZE='442499072'
+fi
+
+# check for old/non-cwm recovery.
+if ! /tmp/busybox test -n "$UPDATE_PACKAGE" ; then
+    # scrape package location from /tmp/recovery.log
+    UPDATE_PACKAGE=`/tmp/busybox cat /tmp/recovery.log | /tmp/busybox grep 'Update location:' | /tmp/busybox tail -n 1 | /tmp/busybox cut -d ' ' -f 3-`
+fi
+
+# check if we're running on a bml, mtd (old) or mtd (current) device
+if /tmp/busybox test -e /dev/block/bml7 ; then
+    # we're running on a bml device
+
     # make sure sdcard is mounted
-    check_mount /mnt/sdcard dev/block/mmcblk0p1 vfat
+    check_mount /mnt/sdcard $SD_PART vfat
 
     # everything is logged into /mnt/sdcard/cyanogenmod_bml.log
     set_log /mnt/sdcard/cyanogenmod_bml.log
 
-    # make sure efs is mounted
-    check_mount /efs /dev/block/stl3 rfs
+    if $IS_GSM ; then
+        # make sure efs is mounted
+        check_mount /efs /dev/block/stl3 rfs
 
-    # create a backup of efs
-    if /tmp/busybox test -e /mnt/sdcard/backup/efs ; then
-        /tmp/busybox mv /mnt/sdcard/backup/efs /mnt/sdcard/backup/efs-$$
+        # create a backup of efs
+        if /tmp/busybox test -e /mnt/sdcard/backup/efs ; then
+            /tmp/busybox mv /mnt/sdcard/backup/efs /mnt/sdcard/backup/efs-$$
+        fi
+        /tmp/busybox rm -rf /mnt/sdcard/backup/efs
+
+        /tmp/busybox mkdir -p /mnt/sdcard/backup/efs
+        /tmp/busybox cp -R /efs/ /mnt/sdcard/backup
     fi
-    /tmp/busybox rm -rf /mnt/sdcard/backup/efs
-    
-    /tmp/busybox mkdir -p /mnt/sdcard/backup/efs
-    /tmp/busybox cp -R /efs/ /mnt/sdcard/backup
 
     # write the package path to sdcard cyanogenmod.cfg
     if /tmp/busybox test -n "$UPDATE_PACKAGE" ; then
-        /tmp/busybox echo `fix_package_location $UPDATE_PACKAGE` > /sdcard/cyanogenmod.cfg
+        /tmp/busybox echo `fix_package_location $UPDATE_PACKAGE` > /mnt/sdcard/cyanogenmod.cfg
     fi
 
     # Scorch any ROM Manager settings to require the user to reflash recovery
@@ -84,72 +147,126 @@ fix_package_location() {
     if [ "$?" != "0" ] ; then
         exit 3
     fi
+    /tmp/busybox sync
+
+    /tmp/busybox reboot now
+    exit 0
+
+elif /tmp/busybox test `/tmp/busybox cat /sys/class/mtd/mtd2/name` != "datadata" ; then
+    # we're running on a mtd (old) device
+
+    # make sure sdcard is mounted
+    check_mount /sdcard $SD_PART vfat
+
+    # everything is logged into /sdcard/cyanogenmod_mtd_old.log
+    set_log /sdcard/cyanogenmod_mtd_old.log
+
+    warn_repartition
+
+    # write the package path to sdcard cyanogenmod.cfg
+    if /tmp/busybox test -n "$UPDATE_PACKAGE" ; then
+        /tmp/busybox echo `fix_package_location $UPDATE_PACKAGE` > /sdcard/cyanogenmod.cfg
+    fi
+
+    # inform the script that this is an old mtd upgrade
+    /tmp/busybox echo 1 > /sdcard/cyanogenmod.mtdupd
+
+    # clear datadata
+    /tmp/busybox umount -l /datadata
+    /tmp/erase_image datadata
+
+    # write new kernel to boot partition
+    /tmp/bml_over_mtd.sh boot 72 reservoir 2004 /tmp/boot.img
+
+	# Remove /system/build.prop to trigger emergency boot
+	/tmp/busybox mount /system
+	/tmp/busybox rm -f /system/build.prop
+	/tmp/busybox umount -l /system
 
     /tmp/busybox sync
 
     /tmp/busybox reboot now
-    exit 9
+    exit 0
+
 elif /tmp/busybox test -e /dev/block/mtdblock0 ; then
-# we're running on a mtd device
+    # we're running on a mtd (current) device
 
     # make sure sdcard is mounted
-    check_mount /sdcard dev/block/mmcblk0p1 vfat
-
-    # remove old log
-    rm -rf /sdcard/cyanogenmod_mtd.log
+    check_mount /sdcard $SD_PART vfat
 
     # everything is logged into /sdcard/cyanogenmod.log
     set_log /sdcard/cyanogenmod_mtd.log
 
-    # if a cyanogenmod.cfg exists, then this is a first time install
-    # let's format the volumes and restore efs
+    # unmount system and data (recovery seems to expect system to be unmounted)
+    /tmp/busybox umount -l /system
+    /tmp/busybox umount -l /data
+
+    # Resize partitions
+    # (For first install, this will get skipped because device doesn't exist)
+    if /tmp/busybox test `/tmp/busybox blockdev --getsize64 /dev/mapper/lvpool-system` -lt $SYSTEM_SIZE ; then
+        warn_repartition
+        /lvm/sbin/lvm lvremove -f lvpool
+        format_partitions
+    fi
+
     if ! /tmp/busybox test -e /sdcard/cyanogenmod.cfg ; then
+        # update install - flash boot image then skip back to updater-script
+        # (boot image is already flashed for first time install or old mtd upgrade)
+
+        # flash boot image
+        /tmp/bml_over_mtd.sh boot 72 reservoir 2004 /tmp/boot.img
+
+        if ! $IS_GSM ; then
+            /tmp/bml_over_mtd.sh recovery 102 reservoir 2004 /tmp/recovery.bin
+        fi
+
         exit 0
     fi
-	
+
+    # if a cyanogenmod.cfg exists, then this is a first time install
+    # let's format the volumes and restore efs
+
     # remove the cyanogenmod.cfg to prevent this from looping
     /tmp/busybox rm -f /sdcard/cyanogenmod.cfg
 
-    # unmount, format and mount system
-    /tmp/busybox umount -l /system
-    /tmp/erase_image system
-    /tmp/busybox mount -t yaffs2 /dev/block/mtdblock2 /system
+    # setup lvm volumes
+    /lvm/sbin/lvm pvcreate $MMC_PART
+    /lvm/sbin/lvm vgcreate lvpool $MMC_PART
+    format_partitions
 
-    # unmount and format cache
-    /tmp/busybox umount -l /cache
-    /tmp/erase_image cache
+    # restart into recovery so the user can install further packages before booting
+    /tmp/busybox touch /cache/.startrecovery
 
-    # unmount and format data
-    /tmp/busybox umount /data
-    /tmp/make_ext4fs -b 4096 -g 32768 -i 8192 -I 256 -a /data /dev/block/mmcblk0p2
+    if /tmp/busybox test -e /sdcard/cyanogenmod.mtdupd ; then
+        # this is an upgrade with changed MTD mapping for /data, /cache, /system
+        # so return to updater-script after formatting them
 
-    # unmount and format datadata
-    /tmp/busybox umount -l /datadata
-    /tmp/erase_image dbdatafs
+        /tmp/busybox rm -f /sdcard/cyanogenmod.mtdupd
 
-    # restore efs backup
-    if /tmp/busybox test -e /sdcard/backup/efs/serial.info ; then
-        /tmp/busybox umount -l /efs
-        /tmp/erase_image efs
-        /tmp/busybox mkdir -p /efs
-
-        if ! /tmp/busybox grep -q /efs /proc/mounts ; then
-            if ! /tmp/busybox mount -t yaffs2 /dev/block/mtdblock4 /efs ; then
-                /tmp/busybox echo "Cannot mount efs."
-                exit 6
-            fi
-        fi
-
-        /tmp/busybox cp -R /sdcard/backup/efs /
-        /tmp/busybox umount -l /efs
-    else
-        /tmp/busybox echo "Cannot restore efs."
-        exit 7
+        exit 0
     fi
 
-    # flash boot image
-    /tmp/bml_over_mtd.sh boot 72 reservoir 2004 /tmp/boot.img
-    /tmp/bml_over_mtd.sh recovery 102 reservoir 2004 /tmp/recovery.bin
+    if $IS_GSM ; then
+        # restore efs backup
+        if /tmp/busybox test -e /sdcard/backup/efs/serial.info ; then
+            /tmp/busybox umount -l /efs
+            /tmp/erase_image efs
+            /tmp/busybox mkdir -p /efs
+
+            if ! /tmp/busybox grep -q /efs /proc/mounts ; then
+                if ! /tmp/busybox mount -t yaffs2 /dev/block/mtdblock4 /efs ; then
+                    /tmp/busybox echo "Cannot mount efs."
+                    exit 6
+                fi
+            fi
+
+            /tmp/busybox cp -R /sdcard/backup/efs /
+            /tmp/busybox umount -l /efs
+        else
+            /tmp/busybox echo "Cannot restore efs."
+            exit 7
+        fi
+    fi
 
     exit 0
 fi
